Generalise configuration reading with a trait
